# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "AddBatchShipmentItemRequest": {
        "properties": {
          "dev_eui": {
            "title": "Dev Eui",
            "type": "string"
          }
        },
        "required": [
          "dev_eui"
        ],
        "title": "AddBatchShipmentItemRequest",
        "type": "object"
      },
...
      "BatchListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "BatchListResponse",
        "type": "object"
      },
...
      "BatchResponse": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "day_plan_qty": {
            "title": "Day Plan Qty",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dev_eui_prefix": {
            "$ref": "#/components/schemas/KgDevEuiPrefixSummaryResponse"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kg_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KgVersionSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "planned_qty": {
            "title": "Planned Qty",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "planned_qty",
          "day_plan_qty",
          "status",
          "dev_eui_prefix",
          "kg_version",
          "created_by_user_id",
          "created_by_user",
          "created_at",
          "updated_at",
          "completed_at",
          "archived_at"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
      "BatchShipmentItemResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kg_dev_eui": {
            "title": "Kg Dev Eui",
            "type": "string"
          },
          "shipment_id": {
            "format": "uuid",
            "title": "Shipment Id",
            "type": "string"
          }
        },
        "required": [
          "shipment_id",
          "kg_dev_eui",
          "created_at"
        ],
        "title": "BatchShipmentItemResponse",
        "type": "object"
      },
      "BatchShipmentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchShipmentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "BatchShipmentListResponse",
        "type": "object"
      },
      "BatchShipmentResponse": {
        "properties": {
          "batch_id": {
            "format": "uuid",
            "title": "Batch Id",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "void_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Void Reason"
          },
          "voided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided At"
          }
        },
        "required": [
          "id",
          "batch_id",
          "comment",
          "quantity",
          "created_by_user_id",
          "created_by_user",
          "created_at",
          "updated_at",
          "completed_at",
          "voided_at",
          "void_reason"
        ],
        "title": "BatchShipmentResponse",
        "type": "object"
      },
      "BatchStatus": {
        "enum": [
          "IN_PRODUCTION",
          "COMPLETED"
        ],
        "title": "BatchStatus",
        "type": "string"
      },
...
      "CreateBatchRequest": {
        "properties": {
          "day_plan_qty": {
            "exclusiveMinimum": 0.0,
            "title": "Day Plan Qty",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dev_eui_prefix": {
            "title": "Dev Eui Prefix",
            "type": "string"
          },
          "kg_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kg Version Id"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "planned_qty": {
            "exclusiveMinimum": 0.0,
            "title": "Planned Qty",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "dev_eui_prefix",
          "planned_qty",
          "day_plan_qty"
        ],
        "title": "CreateBatchRequest",
        "type": "object"
      },
      "CreateBatchShipmentRequest": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "title": "CreateBatchShipmentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KgDevEuiPrefixSummaryResponse": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "short_code": {
            "title": "Short Code",
            "type": "string"
          }
        },
        "required": [
          "prefix",
          "short_code",
          "name"
        ],
        "title": "KgDevEuiPrefixSummaryResponse",
        "type": "object"
      },
...
      "KgVersionSummaryResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "KgVersionSummaryResponse",
        "type": "object"
      },
...
      "UpdateBatchShipmentRequest": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "title": "UpdateBatchShipmentRequest",
        "type": "object"
      },
...
      "UserSummaryResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UserSummaryResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoidBatchShipmentRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "VoidBatchShipmentRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "backend",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/batches": {
      "post": {
        "operationId": "batch-create_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch",
        "tags": [
          "batch"
        ]
      }
    },
    "/batches/": {
      "get": {
        "operationId": "batch-list_batches",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BatchStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at",
              "enum": [
                "name",
                "planned_qty",
                "day_plan_qty",
                "status",
                "created_at",
                "updated_at",
                "completed_at",
                "archived_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batches",
        "tags": [
          "batch"
        ]
      }
    },
...
    "/batches/{batch_id}/shipments": {
      "get": {
        "operationId": "batch-list_batch_shipments",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_voided",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Voided",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchShipmentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batch Shipments",
        "tags": [
          "batch"
        ]
      },
      "post": {
        "operationId": "batch-create_batch_shipment",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchShipmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchShipmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch Shipment",
        "tags": [
          "batch"
        ]
      }
    },
    "/batches/{batch_id}/shipments/{shipment_id}": {
      "patch": {
        "operationId": "batch-update_batch_shipment",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shipment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBatchShipmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchShipmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Batch Shipment",
        "tags": [
          "batch"
        ]
      }
    },
    "/batches/{batch_id}/shipments/{shipment_id}/complete": {
      "post": {
        "operationId": "batch-complete_batch_shipment",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shipment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchShipmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Batch Shipment",
        "tags": [
          "batch"
        ]
      }
    },
    "/batches/{batch_id}/shipments/{shipment_id}/items": {
      "get": {
        "operationId": "batch-list_batch_shipment_items",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shipment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchShipmentItemResponse"
                  },
                  "title": "Response Batch-List Batch Shipment Items",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Batch Shipment Items",
        "tags": [
          "batch"
        ]
      },
      "post": {
        "operationId": "batch-add_batch_shipment_item",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shipment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBatchShipmentItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchShipmentItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Batch Shipment Item",
        "tags": [
          "batch"
        ]
      }
    },
    "/batches/{batch_id}/shipments/{shipment_id}/items/{dev_eui}": {
      "delete": {
        "operationId": "batch-remove_batch_shipment_item",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shipment Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dev_eui",
            "required": true,
            "schema": {
              "title": "Dev Eui",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Batch Shipment Item",
        "tags": [
          "batch"
        ]
      }
    },
    "/batches/{batch_id}/shipments/{shipment_id}/void": {
      "post": {
        "operationId": "batch-void_batch_shipment",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shipment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shipment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidBatchShipmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchShipmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Void Batch Shipment",
        "tags": [
          "batch"
        ]
      }
    },
...
  }
}